Updates the python tests to produce the same test explorer structure at the windows tests.
